Add files via upload
Chatbots work by using artificial intelligence (AI) and natural language processing (NLP) technologies to understand and interpret human language. When a user interacts with a chatbot, it analyzes the input and tries to understand its intent.
A chatbot is designed to work without the assistance of a human operator. AI chatbot responds to questions posed to it in natural language as if it were a real person. It responds using a combination of pre-programmed scripts and machine learning algorithms.